Move ModelFolderObserver into separate class file
